media/requirements.txt
quart
ujson
uvloop
asyncpg
surrealdb
quart-schema
quart-redis
quart-jwt-extended
google-cloud-storage